refactor: restructure search
1. remove dataset / deps related functionalities.
2. add document search index / apis.
