+++ crates/tabby-common/src/api/mod.rs
@@ -1,4 +1,5 @@
 pub mod code;
+pub mod doc;
 pub mod event;
 pub mod server_setting;
