+++ crates/tabby-common/src/index/code.rs
