+++ crates/tabby-common/src/index/mod.rs
@@ -1,0 +1,5 @@
+mod code;
+pub use code::{register_tokenizers, CodeSearchSchema};
+
+mod doc;
+pub use doc::DocSearchSchema;
