+++ crates/tabby-scheduler/Cargo.toml
@@ -12,8 +12,6 @@ tantivy = { workspace = true }
 tracing = { workspace = true }
 tree-sitter-tags = "0.20.2"
 lazy_static = { workspace = true }
-serde-jsonlines = { workspace = true }
-file-rotate = "0.7.5"
 tree-sitter-python = "0.20.2"
 tree-sitter-java = "0.20.2"
 tree-sitter-kotlin = "0.3.1"
@@ -26,14 +24,8 @@ tree-sitter-cpp = { git = "https://github.com/tree-sitter/tree-sitter-cpp", rev 
 tree-sitter-c-sharp = "0.20.0"
 tree-sitter-solidity = { git = "https://github.com/JoranHonig/tree-sitter-solidity", rev = "b239a95" }
 ignore.workspace = true
-requirements = "0.3.0"
-serdeconv.workspace = true
-cargo-lock = { version = "9.0.0", features = ["dependency-tree"] }
 tokio-cron-scheduler = { workspace = true }
 tokio = { workspace = true, features = ["process"] }
-package-lock-json-parser = "0.4.0"
-npm-package-json = "0.1.3"
-yarn-lock-parser = "0.7.0"
 text-splitter = "0.10.0"
 kv = { version = "0.24.0", features = ["json-value"] }
 serde.workspace = true
