+++ crates/tabby-scheduler/src/code/mod.rs
@@ -1,8 +1,13 @@
-use tabby_common::{Point, Tag};
-use text_splitter::{Characters, TextSplitter};
-use tree_sitter_tags::TagsContext;
-
 mod languages;
+mod types;
+
+use std::{fs::read_to_string, path::Path};
+
+use tabby_common::{config::RepositoryConfig, languages::get_language_by_ext};
+use text_splitter::{Characters, TextSplitter};
+use tracing::warn;
+use tree_sitter_tags::TagsContext;
+pub use types::{Point, SourceFile, Tag};
 
 pub struct CodeIntelligence {
     context: TagsContext,
@@ -53,11 +58,88 @@ impl CodeIntelligence {
             .collect()
     }
 
+    pub fn create_source_file(
+        &mut self,
+        config: &RepositoryConfig,
+        path: &Path,
+    ) -> Option<SourceFile> {
+        if path.is_dir() || !path.exists() {
+            return None;
+        }
+        let relative_path = path
+            .strip_prefix(&config.dir())
+            .expect("Paths always begin with the prefix");
+
+        let Some(ext) = relative_path.extension() else {
+            return None;
+        };
+
+        let Some(language_info) = get_language_by_ext(ext) else {
+            warn!("Unknown language for extension {:?}", ext);
+            return None;
+        };
+
+        let language = language_info.language();
+        let contents = match read_to_string(path) {
+            Ok(x) => x,
+            Err(_) => {
+                warn!("Failed to read {path:?}, skipping...");
+                return None;
+            }
+        };
+        let source_file = SourceFile {
+            git_url: config.canonical_git_url(),
+            basedir: config.dir().display().to_string(),
+            filepath: relative_path.display().to_string(),
+            max_line_length: metrics::max_line_length(&contents),
+            avg_line_length: metrics::avg_line_length(&contents),
+            alphanum_fraction: metrics::alphanum_fraction(&contents),
+            tags: self.find_tags(language, &contents),
+            language: language.into(),
+        };
+        Some(source_file)
+    }
+
     // FIXME(meng): implement with treesitter based CodeSplitter.
     pub fn chunks<'splitter, 'text: 'splitter>(
         &'splitter self,
         text: &'text str,
     ) -> impl Iterator<Item = &'text str> + 'splitter {
         self.splitter.chunks(text, 192)
+    }
+}
+
+mod metrics {
+    use std::cmp::max;
+
+    pub fn max_line_length(content: &str) -> usize {
+        content.lines().map(|x| x.len()).reduce(max).unwrap_or(0)
+    }
+
+    pub fn avg_line_length(content: &str) -> f32 {
+        let mut total = 0;
+        let mut len = 0;
+        for x in content.lines() {
+            len += 1;
+            total += x.len();
+        }
+
+        if len > 0 {
+            total as f32 / len as f32
+        } else {
+            0.0
+        }
+    }
+
+    pub fn alphanum_fraction(content: &str) -> f32 {
+        let num_alphanumn: f32 = content
+            .chars()
+            .map(|x| f32::from(u8::from(x.is_alphanumeric())))
+            .sum();
+        if !content.is_empty() {
+            num_alphanumn / content.len() as f32
+        } else {
+            0.0
+        }
     }
 }
