+++ crates/tabby-scheduler/src/code/types.rs
@@ -1,0 +1,47 @@
+use std::{ops::Range, path::Path};
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct SourceFile {
+    pub git_url: String,
+    pub basedir: String,
+    pub filepath: String,
+    pub language: String,
+    pub max_line_length: usize,
+    pub avg_line_length: f32,
+    pub alphanum_fraction: f32,
+    pub tags: Vec<Tag>,
+}
+
+impl SourceFile {
+    pub fn read_content(&self) -> std::io::Result<String> {
+        let path = Path::new(&self.basedir).join(&self.filepath);
+        std::fs::read_to_string(path)
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct Point {
+    pub row: usize,
+    pub column: usize,
+}
+
+impl Point {
+    pub fn new(row: usize, column: usize) -> Self {
+        Self { row, column }
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct Tag {
+    pub range: Range<usize>,
+    pub name_range: Range<usize>,
+    pub utf16_column_range: Range<usize>,
+    pub span: Range<Point>,
+    pub line_range: Range<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub docs: Option<String>,
+    pub is_definition: bool,
+    pub syntax_type_name: String,
+}
